# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: RoomFlow API
  version: 1.0.0
  contact:
    name: RoomFlow Team
  description: |-
    RoomFlow 会议室预约管理系统 REST API 契约（OpenAPI 3.0.3）。

    本文件是前后端唯一共享契约：后端 Controller 与前端 api/types 均以此为准。
    配套说明文档（错误码表、分页约定、权限矩阵、时区约定）见 `docs/api/README.md`。

    通用约定：
    - 统一响应体 `Result<T>`：`{code:int, message:string, data:T}`。`code=0` 成功；非 0 为业务错误码，同时携带匹配的 HTTP 状态码。
    - 时间：全部时间字段（请求与响应）均为北京时间，ISO 8601 格式且必须显式携带 `+08:00` 偏移，例：`2026-09-16T14:00:00+08:00`。`format: date` 的查询参数为北京时间日历日 `YYYY-MM-DD`。
    - 会议时段为 `[startTime, endTime)` 半开区间；同一房间的相邻时段允许首尾相接（前一会的 endTime 可等于后一会的 startTime）。
    - 时间规则：`startTime` 不得早于当前时间、不得晚于当前时间 +72h（滚动窗口只约束开始时间）；`startTime` 与 `endTime` 均须按 15 分钟对齐（秒=0 且分钟 %15=0）；时长为 15min–24h（含两端）；允许跨天；`endTime` 可超出 72h 预约窗口。
    - 权限：每个端点以 `x-permission` 标注，取值 `PUBLIC`（无需登录）/ `REFRESH_TOKEN`（仅凭请求体中的 Refresh Token）/ `AUTHENTICATED`（任意已登录用户）/ `ADMIN` / `ORGANIZER_OR_ADMIN`（会议发起人或管理员）/ `SELF`（仅操作本人资源）。全局默认要求 Bearer JWT；公开端点显式声明 `security: []`。
    - 每个端点可能返回的业务错误码以 `x-error-codes` 枚举，与 README 错误码表一一对应。
    - 分页：`page`（从 1 起）/ `size`（默认 10，最大 100），分页响应为 `PageResult<T> = {records, total, page, size}`。
servers:
  - url: http://localhost:8080
    description: 本地开发环境
...
security:
  - bearerAuth: []
paths:
...
  /api/meetings/{id}/participants:
    post:
      tags:
        - participants
      operationId: joinMeeting
      summary: 报名参会
      description: |-
        当前登录用户报名加入会议。无请求体。
        服务端按顺序校验（任一失败即返回对应错误码）：
        1. 会议存在且 status 非 DELETED → 否则 40401；
        2. 会议 status=ACTIVE 且未开始（startTime > now）→ 否则 40909（已开始的会议停止报名）；
        3. 本人无 banned=1（leaveReason=KICKED）记录 → 否则 40904（被踢永久禁入，不重置）；
        4. 本人无有效参会记录（leftAt 为 null，含发起人记录）→ 否则 40908；
        5. 有效参会人数 < 房间 capacity → 否则 40903；
        6. 存在 leaveReason=USER_LEFT 的历史记录时复用该行（leftAt/leaveReason 置 null，banned 保持 0），
           不新增行，同样返回 200。
        成功后向发起人发送报名通知（异步，不在本响应中体现）。
      x-permission: AUTHENTICATED
      x-error-codes: [40001, 40101, 40102, 40401, 40903, 40904, 40908, 40909, 50000]
      parameters:
        - $ref: '#/components/parameters/PathId'
      responses:
        '200':
          description: 报名成功（含退出后重新报名）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultParticipantVO'
              example:
                code: 0
                message: success
                data:
                  id: 503
                  meetingId: 101
                  accountId: 1003
                  username: carol
                  isOrganizer: false
                  banned: false
                  joinedAt: '2026-09-15T20:00:00+08:00'
                  leftAt: null
                  leaveReason: null
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalError'
    get:
      tags:
        - participants
      operationId: listParticipants
      summary: 参会者列表
      description: |-
        返回该会议的全部参会记录（含已退出与被踢记录），按 joinedAt 升序、id 升序。
        有效参会者 = leftAt 为 null 的记录；participantCount 口径与此一致。
        会议不存在或 status=DELETED 返回 40401。
      x-permission: AUTHENTICATED
      x-error-codes: [40001, 40101, 40102, 40401, 50000]
      parameters:
        - $ref: '#/components/parameters/PathId'
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultParticipantVOList'
              example:
                code: 0
                message: success
                data:
                  - id: 501
                    meetingId: 101
                    accountId: 1001
                    username: alice
                    isOrganizer: true
                    banned: false
                    joinedAt: '2026-09-15T16:30:00+08:00'
                    leftAt: null
                    leaveReason: null
                  - id: 502
                    meetingId: 101
                    accountId: 1002
                    username: bob
                    isOrganizer: false
                    banned: false
                    joinedAt: '2026-09-15T17:00:00+08:00'
                    leftAt: '2026-09-15T18:00:00+08:00'
                    leaveReason: USER_LEFT
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/meetings/{id}/participants/me:
    delete:
      tags:
        - participants
      operationId: leaveMeeting
      summary: 退出报名（本人）
      description: |-
        当前登录用户退出自己的有效参会记录：leftAt=now、leaveReason=USER_LEFT，banned 保持 0。
        会议须为 ACTIVE 且未开始（已开始/已结束/已取消的会议无需退出，返回 40909）。
        发起人不能退出自己的会议（应使用 cancel），返回 40909。
        本人无有效参会记录返回 40401。退出后允许重新报名。无请求体。
        成功后向发起人发送退出通知（异步）。
      x-permission: SELF
      x-error-codes: [40001, 40101, 40102, 40401, 40909, 50000]
      parameters:
        - $ref: '#/components/parameters/PathId'
      responses:
        '200':
          description: 退出成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultVoid'
              example:
                code: 0
                message: success
                data: null
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/meetings/{id}/participants/{accountId}:
    delete:
      tags:
        - participants
      operationId: kickParticipant
      summary: 踢出参会者
      description: |-
        将指定账号移出会议并永久禁止再次加入：banned=1、leaveReason=KICKED、leftAt=now。
        仅发起人或 ADMIN。会议须为 ACTIVE，否则 40909。
        目标账号须存在有效参会记录（leftAt 为 null），否则 40401；
        不能踢发起人（accountId=organizerId），返回 40909。
        成功后向被踢人发送踢人通知（异步）。无请求体。
      x-permission: ORGANIZER_OR_ADMIN
      x-error-codes: [40001, 40101, 40102, 40301, 40401, 40909, 50000]
      parameters:
        - $ref: '#/components/parameters/PathId'
        - $ref: '#/components/parameters/PathAccountId'
      responses:
        '200':
          description: 踢人成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultVoid'
              example:
                code: 0
                message: success
                data: null
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/notifications:
    get:
      tags:
        - notifications
      operationId: listNotifications
      summary: 站内通知列表
      description: |-
        分页查询当前登录用户的站内通知。
        默认排序：未读优先（isRead ASC），同组内按 createdAt DESC、id DESC。
        可选 `isRead` 筛选；`isRead=false&size=1` 可用于获取未读总数（total 即未读数）。
      x-permission: AUTHENTICATED
      x-error-codes: [40001, 40101, 40102, 50000]
      parameters:
        - $ref: '#/components/parameters/QueryPage'
        - $ref: '#/components/parameters/QuerySize'
        - name: isRead
          in: query
          required: false
          schema:
            type: boolean
          description: 按已读状态筛选；缺省返回全部
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultPageNotificationVO'
              example:
                code: 0
                message: success
                data:
                  records:
                    - id: 9001
                      type: PARTICIPANT_JOINED
                      title: 新参会者
                      content: 用户 bob 报名了你的会议「产品评审会」。
                      meetingId: 101
                      isRead: false
                      createdAt: '2026-09-15T17:00:00+08:00'
                    - id: 9000
                      type: MEETING_ENDED
                      title: 会议已结束
                      content: '会议「周会」已于 2026-09-15 12:00 结束。'
                      meetingId: 99
                      isRead: true
                      createdAt: '2026-09-15T12:00:00+08:00'
                  total: 2
                  page: 1
                  size: 10
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/notifications/{id}/read:
    patch:
      tags:
        - notifications
      operationId: markNotificationRead
      summary: 标记单条已读
      description: |-
        将当前登录用户的指定通知标记为已读。幂等：已读的通知重复调用仍返回 200。
        通知不存在或不属于当前用户均返回 40401（不区分，防止枚举他人通知 ID）。无请求体。
      x-permission: SELF
      x-error-codes: [40001, 40101, 40102, 40401, 50000]
      parameters:
        - $ref: '#/components/parameters/PathId'
      responses:
        '200':
          description: 标记成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultNotificationVO'
              example:
                code: 0
                message: success
                data:
                  id: 9001
                  type: PARTICIPANT_JOINED
                  title: 新参会者
                  content: 用户 bob 报名了你的会议「产品评审会」。
                  meetingId: 101
                  isRead: true
                  createdAt: '2026-09-15T17:00:00+08:00'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/notifications/read-all:
    patch:
      tags:
        - notifications
      operationId: markAllNotificationsRead
      summary: 全部标记已读
      description: |-
        将当前登录用户的全部未读通知标记为已读，返回本次更新的条数（data 为整数）。
        路径常量 `read-all` 与 `/{id}` 不冲突：{id} 仅匹配 int64。无请求体。
      x-permission: AUTHENTICATED
      x-error-codes: [40101, 40102, 50000]
      responses:
        '200':
          description: 标记成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultInt'
              example:
                code: 0
                message: success
                data: 3
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |-
        将 Access Token 放入请求头：`Authorization: Bearer <accessToken>`。
        未携带 / 签名无效 / 已被撤销（logout 或账号禁用后）→ 401 + code=40101；
        已过期 → 401 + code=40102（前端应调用 POST /api/auth/refresh 换新）。
  parameters:
    PathId:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
        minimum: 1
      description: 资源 ID
    PathAccountId:
      name: accountId
      in: path
      required: true
      schema:
        type: integer
        format: int64
        minimum: 1
      description: 账号 ID
    QueryPage:
      name: page
      in: query
      required: false
      schema:
        type: integer
        format: int32
        minimum: 1
        default: 1
      description: 页码，从 1 开始
    QuerySize:
      name: size
      in: query
      required: false
      schema:
        type: integer
        format: int32
        minimum: 1
        maximum: 100
        default: 10
      description: 每页条数，最大 100
  responses:
    BadRequest:
      description: 参数校验失败（code=40001，data.fieldErrors 给出字段级错误）或请求格式错误（code=40002，如 JSON 解析失败 / 类型不匹配）
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResult'
          example:
            code: 40001
            message: 参数校验失败
            data:
              fieldErrors:
                - field: startTime
                  message: 必须按 15 分钟对齐
    Unauthorized:
      description: 未认证或凭证无效（code=40101 未携带/无效/已撤销，40102 Token 过期，40103 用户名或密码错误，40104 Refresh Token 无效/过期/已撤销）
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResult'
          example:
            code: 40101
            message: 未认证或 Token 无效
            data: null
    Forbidden:
      description: 已认证但无权限（code=40301 无权限，40302 账号已禁用）
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResult'
          example:
            code: 40301
            message: 无权限执行该操作
            data: null
    NotFound:
      description: 资源不存在或对当前用户不可见（code=40401）
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResult'
          example:
            code: 40401
            message: 资源不存在
            data: null
    Conflict:
      description: 业务规则冲突，code 取值见该端点 x-error-codes（40901–40909）与 README 错误码表
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResult'
          example:
            code: 40902
            message: 会议时间与该房间已有会议冲突
            data: null
    InternalError:
      description: 服务器内部错误（code=50000），响应不包含堆栈或内部细节
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResult'
          example:
            code: 50000
            message: 服务器内部错误
            data: null
  schemas:
    FieldError:
      type: object
      required:
        - field
        - message
      properties:
        field:
          type: string
          description: 校验失败的字段名（JSON 字段路径）
        message:
          type: string
          description: 字段级错误信息（中文）
    ErrorResult:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
          format: int32
          description: 非 0 业务错误码，见 README 错误码表
        message:
          type: string
          description: 人类可读错误信息（中文），不含敏感内部细节
        data:
          type: object
          nullable: true
          description: 通常为 null；code=40001 时携带 fieldErrors 字段级错误列表
          properties:
            fieldErrors:
              type: array
              items:
                $ref: '#/components/schemas/FieldError'
    Result:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
          format: int32
          description: 0=成功
        message:
          type: string
          description: 提示信息（中文），成功时为 success 或操作描述
        data:
          description: 业务数据，类型见各 ResultXxx 包装 schema；无数据操作时为 null
    ResultVoid:
      allOf:
        - $ref: '#/components/schemas/Result'
        - type: object
          properties:
            data:
              type: object
              nullable: true
              description: 固定为 null
    ResultInt:
      allOf:
        - $ref: '#/components/schemas/Result'
        - type: object
          properties:
            data:
              type: integer
              format: int32
              description: 整数结果（如批量操作影响条数）
...
    ResultParticipantVO:
      allOf:
        - $ref: '#/components/schemas/Result'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/ParticipantVO'
    ResultParticipantVOList:
      allOf:
        - $ref: '#/components/schemas/Result'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/ParticipantVO'
    ResultNotificationVO:
      allOf:
        - $ref: '#/components/schemas/Result'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/NotificationVO'
    ResultPageNotificationVO:
      allOf:
        - $ref: '#/components/schemas/Result'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/PageNotificationVO'
...
    PageNotificationVO:
      type: object
      required:
        - records
        - total
        - page
        - size
      properties:
        records:
          type: array
          items:
            $ref: '#/components/schemas/NotificationVO'
        total:
          type: integer
          format: int64
        page:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
...
    ParticipantVO:
      type: object
      required:
        - id
        - meetingId
        - accountId
        - username
        - isOrganizer
        - banned
        - joinedAt
        - leftAt
        - leaveReason
      properties:
        id:
          type: integer
          format: int64
        meetingId:
          type: integer
          format: int64
        accountId:
          type: integer
          format: int64
        username:
          type: string
          maxLength: 50
        isOrganizer:
          type: boolean
          description: 是否发起人（创建会议时自动写入 true）
        banned:
          type: boolean
          description: 是否被踢禁入；true 时 leaveReason 必为 KICKED
        joinedAt:
          type: string
          format: date-time
        leftAt:
          type: string
          format: date-time
          nullable: true
          description: null=有效参会中；非 null=已退出/被踢
        leaveReason:
          type: string
          enum:
            - USER_LEFT
            - KICKED
            - null
          nullable: true
          description: null=未退出；USER_LEFT=主动退出（可重新报名）；KICKED=被踢（banned=1，永久禁入）
...
    NotificationType:
      type: string
      enum:
        - PARTICIPANT_JOINED
        - PARTICIPANT_LEFT
        - PARTICIPANT_KICKED
        - MEETING_ENDED
      description: |-
        通知类型与默认收件人（Open Questions 已批准默认值）：
        PARTICIPANT_JOINED=有人报名→发发起人；PARTICIPANT_LEFT=有人退出→发发起人；
        PARTICIPANT_KICKED=被踢→发被踢人；MEETING_ENDED=会议结束（含到期自动结束与提前结束）→发全体有效参会者。
        会议修改/取消/删除不产生通知。
    NotificationVO:
      type: object
      required:
        - id
        - type
        - title
        - content
        - meetingId
        - isRead
        - createdAt
      properties:
        id:
          type: integer
          format: int64
        type:
          $ref: '#/components/schemas/NotificationType'
        title:
          type: string
          maxLength: 200
        content:
          type: string
        meetingId:
          type: integer
          format: int64
          nullable: true
          description: 关联会议 ID，可空
        isRead:
          type: boolean
        createdAt:
          type: string
          format: date-time